Reorganize driver tools
