+++ tests/core/v5_1/conftest.py
@@ -1,23 +1,23 @@
 import pytest
 
-from ddht.tools.driver import Network
+from ddht.tools.driver import Tester
 
 
 @pytest.fixture
-def network():
-    return Network()
+def tester():
+    return Tester()
 
 
 @pytest.fixture
-async def alice(network):
-    return network.node()
+async def alice(tester):
+    return tester.node()
 
 
 @pytest.fixture
-async def bob(network):
-    return network.node()
+async def bob(tester):
+    return tester.node()
 
 
 @pytest.fixture
-async def driver(network, alice, bob):
-    return network.session_pair(alice, bob)
+async def driver(tester, alice, bob):
+    return tester.session_pair(alice, bob)
